Refactor to make the simplest implementation of rANS static
