+++ src/Transforms/Algorithms/EntropyCoders/StaticRANS/StaticRANSEncoder.cs
@@ -2,9 +2,9 @@ using System;
 using System.Collections.Generic;
 using BWDPerf.Interfaces;
 
-namespace BWDPerf.Transforms.Algorithms.EntropyCoders.rANS
+namespace BWDPerf.Transforms.Algorithms.EntropyCoders.StaticRANS
 {
-    public class rANSEncoder<TSymbol> : ICoder<TSymbol[], byte>
+    public class StaticRANSEncoder<TSymbol> : ICoder<TSymbol[], byte>
     {
         public IRANSModel<TSymbol> Model { get; }
         public IConverter<TSymbol> Converter { get; }
@@ -18,7 +18,7 @@ namespace BWDPerf.Transforms.Algorithms.EntropyCoders.rANS
         public const int _logB = 8; // b = 256, so we emit a byte when normalizing
         public const int _bMask = 255; // mask to get the last logB bits
 
-        public rANSEncoder(IRANSModel<TSymbol> model, IConverter<TSymbol> converter)
+        public StaticRANSEncoder(IRANSModel<TSymbol> model, IConverter<TSymbol> converter)
         {
             this.Model = model;
             this.Converter = converter;
@@ -28,14 +28,14 @@ namespace BWDPerf.Transforms.Algorithms.EntropyCoders.rANS
         {
             await foreach (var buffer in input)
             {
-                var list = new Stack<byte>();
                 // Initialize the model
                 var header = InitializeModel(buffer.AsSpan());
                 // Print header (static order0 distribution)
                 foreach (var b in header) yield return b;
+                var stream = new byte[buffer.Length];
+                int pos = stream.Length;
 
                 uint state = _L;
-                Console.WriteLine($"State low: {_L}; high: {_L << _logB}");
                 for (int i = buffer.Length - 1; i >= 0 ; i--)
                 {
                     var symbol = buffer[i];
@@ -46,20 +46,17 @@ namespace BWDPerf.Transforms.Algorithms.EntropyCoders.rANS
                     var state_max = ((_L << _logB) >> n) * freq;
                     while(state >= state_max)
                     {
-                        list.Push((byte) (state & _bMask));
+                        stream[--pos] = (byte) (state & _bMask);
                         state >>= _logB;
                     }
                     state = (uint) (((state / freq) << n) + state % freq + cdf);
                 }
-                Console.WriteLine($"Ending state: {state}");
                 // Output the state at the end. There are optimizations for using log(state) bits, but for now 32 bits is ok
                 var bytes = BitConverter.GetBytes(state);
                 foreach (var b in bytes)
                     yield return b;
-                while (list.Count > 0)
-                {
-                    yield return list.Pop();
-                }
+                for (int i = pos; i < stream.Length; i++)
+                    yield return stream[i];
             }
         }
