+++ tests/StaticRANSTests.cs
@@ -1,9 +1,9 @@
 using System.IO;
 using System.Threading.Tasks;
 using BWDPerf.Architecture;
-using BWDPerf.Transforms.Algorithms.EntropyCoders.rANS;
+using BWDPerf.Transforms.Algorithms.EntropyCoders.StaticRANS;
 using BWDPerf.Transforms.Converters;
-using BWDPerf.Transforms.Models.RANS;
+using BWDPerf.Transforms.Models.Static.RANS;
 using BWDPerf.Transforms.Serializers;
 using BWDPerf.Transforms.Sources;
 using Microsoft.VisualStudio.TestTools.UnitTesting;
@@ -11,17 +11,17 @@ using Microsoft.VisualStudio.TestTools.UnitTesting;
 namespace BWDPerf.Tests
 {
     [TestClass]
-    public class rANSTests
+    public class StaticRANSTests
     {
-        private const string _file = "../../../../data/file.md";
-        private const string _compressedFile = "../../../file.md.rans";
-        private const string _decompressedFile = "../../../file.md";
+        private const string _file = "../../../../data/enwik4";
+        private const string _compressedFile = "../../../enwik4.rans";
+        private const string _decompressedFile = "../../../decompressed";
 
         [TestMethod]
         public async Task TestCompression()
         {
             var compressTask = new BufferedFileSource(_file, 10_000_000) // 10MB
-                .ToCoder(new rANSEncoder<byte>(new StaticOrder0<byte>(), new ByteConverter()))
+                .ToCoder(new StaticRANSEncoder<byte>(new StaticOrder0<byte>(), new ByteConverter()))
                 .Serialize(new SerializeToFile(_compressedFile));
 
             await compressTask;
@@ -31,7 +31,7 @@ namespace BWDPerf.Tests
         public async Task TestDecompression()
         {
             var decompressTask = new FileSource(_compressedFile)
-                .ToDecoder(new rANSDecoder<byte>(new StaticOrder0<byte>(), new ByteConverter()))
+                .ToDecoder(new StaticRANSDecoder<byte>(new StaticOrder0<byte>(), new ByteConverter()))
                 .Serialize(new SerializeToFile(_decompressedFile));
 
             await decompressTask;
